requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-celery-beat==2.4.0
django-debug-toolbar==3.7.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
image==1.5.33
kombu==5.2.4
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.33
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
qrcode==7.3.1
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
style==1.1.0
tzdata==2022.6
update==0.0.1
vine==5.0.0
wcwidth==0.2.5
